refactor(FON-468): migrate secretariat-general docs to feature-first
