+++ apps/client/src/features/agenda/context/AgendaContext.tsx
