+++ apps/client/src/features/documents/hooks/useSelection.ts
