+++ apps/client/src/pages/secretariat-general/docs/presentations/PresentationsLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router';
 
-import { PresentationPlanProvider } from './contexts/PresentationPlanProvider';
+import { PresentationPlanProvider } from '@/features/presentations/context/PresentationPlanProvider';
 
 export function PresentationsLayout() {
   return (
